Factored DataProtobufGenerator out of specs.
Further reduces the number of module dependencies in the specs
module that is shared with the API service.
